+++ modules/AWSPowerShell/Cmdlets/CloudTrail/Basic/Find-CTEvent-Cmdlet.cs
@@ -138,9 +138,9 @@ namespace Amazon.PowerShell.Cmdlets.CT
         /// <summary>
         /// <para>
         /// <para>The token to use to get the next page of results after a previous API call. This token
-        /// must be passed in with the same parameters that were specified in the the original
-        /// call. For example, if the original call specified an AttributeKey of 'Username' with
-        /// a value of 'root', the call with NextToken should include those same parameters.</para>
+        /// must be passed in with the same parameters that were specified in the original call.
+        /// For example, if the original call specified an AttributeKey of 'Username' with a value
+        /// of 'root', the call with NextToken should include those same parameters.</para>
         /// </para>
         /// <para>
         /// <br/><b>Note:</b> This parameter is only used if you are manually controlling output pagination of the service API call.