+++ modules/AWSPowerShell/Cmdlets/Redshift/Basic/Restore-RSFromClusterSnapshot-Cmdlet.cs
@@ -79,9 +79,8 @@ namespace Amazon.PowerShell.Cmdlets.RS
         #region Parameter AquaConfigurationStatus
         /// <summary>
         /// <para>
-        /// <para>The value represents how the cluster is configured to use AQUA (Advanced Query Accelerator)
-        /// after the cluster is restored. Possible values include the following.</para><ul><li><para>enabled - Use AQUA if it is available for the current Amazon Web Services Region and
-        /// Amazon Redshift node type.</para></li><li><para>disabled - Don't use AQUA. </para></li><li><para>auto - Amazon Redshift determines whether to use AQUA.</para></li></ul>
+        /// <para>This parameter is retired. It does not set the AQUA configuration status. Amazon Redshift
+        /// automatically determines whether to use AQUA (Advanced Query Accelerator).</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -188,8 +187,7 @@ namespace Amazon.PowerShell.Cmdlets.RS
         #region Parameter ElasticIp
         /// <summary>
         /// <para>
-        /// <para>The elastic IP (EIP) address for the cluster. You don't have to specify the EIP for
-        /// a publicly accessible cluster with AvailabilityZoneRelocation turned on.</para>
+        /// <para>The elastic IP (EIP) address for the cluster.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -387,7 +385,8 @@ namespace Amazon.PowerShell.Cmdlets.RS
         /// <summary>
         /// <para>
         /// <para>The Amazon Resource Name (ARN) of the snapshot associated with the message to restore
-        /// from a cluster.</para>
+        /// from a cluster. You can specify this parameter or <code>snapshotIdentifier</code>,
+        /// but not both.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -410,7 +409,8 @@ namespace Amazon.PowerShell.Cmdlets.RS
         /// <summary>
         /// <para>
         /// <para>The name of the snapshot from which to create the new cluster. This parameter isn't
-        /// case sensitive.</para><para>Example: <code>my-snapshot-id</code></para>
+        /// case sensitive. You can specify this parameter or <code>snapshotArn</code>, but not
+        /// both.</para><para>Example: <code>my-snapshot-id</code></para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(Position = 1, ValueFromPipelineByPropertyName = true)]